# contracts/openapi/granete-api.v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Granete Organization Foundation API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "operationId": "refreshSession",
        "responses": {
          "200": {
            "description": "Rotated access and refresh credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        }
      }
    },
...
    "/platform/organizations": {
      "get": {
        "operationId": "listPlatformOrganizations",
        "responses": {
          "200": {
            "description": "Organizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformOrganization"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/organizations": {
      "post": {
        "operationId": "provisionOrganization",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization provisioned atomically",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationProvisioningResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/media:authorize": {
      "post": {
        "operationId": "authorizeMedia",
        "description": "Issue short-lived, resource-scoped read grants for canonical catalog media files. The caller authenticates with a session bearer (organization scope required); each grant is a media_read credential bound to exactly one file of the caller's organization and expires within minutes. Grants never authenticate business APIs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaAuthorizeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grants for every requested file that exists in the caller's organization (missing files are omitted, not erroring, so responses stay enumeration-safe)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAuthorizeResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
...
    },
...
    "responses": {
      "ApiError": {
        "description": "Structured API error",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "requestId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AuthTransport": {
        "type": "string",
        "enum": [
          "web",
          "mobile",
          "sketchup",
          "support"
        ]
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "normalized_email",
          "name",
          "account_status",
          "email_verified_at",
          "last_login_at",
          "platform_admin",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "normalized_email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "account_status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_login_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "platform_admin": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "License": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "plan",
          "status"
        ],
        "properties": {
          "plan": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "OrganizationSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "slug",
          "type",
          "status",
          "license"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "factory",
              "store",
              "dealer"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/OrganizationStatus"
          },
          "license": {
            "$ref": "#/components/schemas/License"
          }
        }
      },
      "Membership": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "organization_id",
          "user_id",
          "status",
          "roles",
          "joined_at",
          "version",
          "organization"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationSummary"
          }
        }
      },
...
      "LoginResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user",
          "license",
          "transport",
          "token",
          "roles",
          "memberships",
          "selection_required"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "license": {
            "$ref": "#/components/schemas/License"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationSummary"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            }
          },
          "selection_required": {
            "type": "boolean"
          },
          "support": {
            "type": "boolean"
          },
          "transport": {
            "$ref": "#/components/schemas/AuthTransport"
          },
          "session_id": {
            "type": "string",
            "description": "Server-side session registry id (#460). Present on tokens of version 5+; absent while a pre-#460 token is exchanged. Becomes required when the #460 SEC-9 gate retires version 4."
          },
          "refresh_token": {
            "type": "string",
            "description": "Opaque single-use refresh credential. Emitted only for web/mobile during SEC-2; SEC-4 moves web storage to a protected cookie."
          },
          "refresh_expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Never later than the auth session absolute expiry."
          }
        }
      },
...
      "PlatformOrganization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "slug",
          "type",
          "status",
          "license_plan",
          "member_count",
          "created_at",
          "updated_at",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "factory",
              "store",
              "dealer"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/OrganizationStatus"
          },
          "license_plan": {
            "type": "string"
          },
          "license_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "parent_organization_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "member_count": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "ProvisionOrganizationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "type",
          "license_plan"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z0-9]([a-z0-9-]{0,62}[a-z0-9])$"
          },
          "type": {
            "$ref": "#/components/schemas/OrganizationType"
          },
          "license_plan": {
            "type": "string",
            "enum": [
              "none",
              "trial",
              "pro"
            ]
          },
          "license_expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "bootstrap_admin_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "clone_catalog_from": {
            "type": "string",
            "format": "uuid"
          },
          "entitlements": {
            "$ref": "#/components/schemas/ProvisionOrganizationEntitlements"
          }
        }
      },
      "OrganizationProvisioningResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organization",
          "readiness"
        ],
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/PlatformOrganization"
          },
          "readiness": {
            "$ref": "#/components/schemas/OrganizationReadiness"
          }
        }
      },
...
      "RefreshRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "refresh_token",
          "transport"
        ],
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 58,
            "maxLength": 128,
            "description": "Opaque single-use bearer credential; never an access JWT."
          },
          "transport": {
            "$ref": "#/components/schemas/AuthTransport"
          }
        }
      },
...
      "MediaAuthorizeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "resources"
        ],
        "description": "Canonical catalog media filenames (server-generated upload names) to authorize for read. Clients request typed resource ids, never arbitrary URLs.",
        "properties": {
          "resources": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "pattern": "^[0-9a-f]{32}\\\\.(jpg|png|webp)$"
            }
          }
        }
      },
...
      "MediaAuthorizeResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "grants"
        ],
        "properties": {
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaGrant"
            }
          }
        }
      }
    }
  }
}